# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EduLearn EdTech Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: trainerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course (trainer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update course (trainer/admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseBody"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course (trainer/admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /courses/{id}/lessons:
    post:
      operationId: createLesson
      tags: [courses]
      summary: Add a lesson to a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /courses/{id}/rate:
    post:
      operationId: rateCourse
      tags: [courses]
      summary: Rate a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateCourseBody"
      responses:
        "200":
          description: Rating submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseRating"
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List enrollments (student gets own, trainer/admin get all)
      responses:
        "200":
          description: List of enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: enroll
      tags: [enrollments]
      summary: Enroll in a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollBody"
      responses:
        "201":
          description: Enrolled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}/progress:
    patch:
      operationId: updateProgress
      tags: [enrollments]
      summary: Update lesson progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgressBody"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
...
components:
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        trainerId:
          type: integer
        trainerName:
          type: string
        thumbnail:
          type: string
          nullable: true
        price:
          type: number
        enrolledCount:
          type: integer
        rating:
          type: number
        ratingCount:
          type: integer
        lessonCount:
          type: integer
        isPublished:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - category
        - trainerId
        - trainerName
        - price
        - enrolledCount
        - rating
        - ratingCount
        - lessonCount
        - isPublished
        - createdAt
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            lessons:
              type: array
              items:
                $ref: "#/components/schemas/Lesson"
            ratings:
              type: array
              items:
                $ref: "#/components/schemas/CourseRating"
    CreateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        thumbnail:
          type: string
        price:
          type: number
        isPublished:
          type: boolean
      required:
        - title
        - description
        - category
        - price
    UpdateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        thumbnail:
          type: string
        price:
          type: number
        isPublished:
          type: boolean
    Lesson:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        videoUrl:
          type: string
          nullable: true
        pdfUrl:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
        orderIndex:
          type: integer
        moduleTitle:
          type: string
          nullable: true
      required:
        - id
        - courseId
        - title
        - orderIndex
    CreateLessonBody:
      type: object
      properties:
        title:
          type: string
        videoUrl:
          type: string
        pdfUrl:
          type: string
        duration:
          type: integer
        moduleTitle:
          type: string
        orderIndex:
          type: integer
      required:
        - title
    CourseRating:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        rating:
          type: integer
        review:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - courseId
        - studentId
        - studentName
        - rating
        - createdAt
    RateCourseBody:
      type: object
      properties:
        rating:
          type: integer
        review:
          type: string
      required:
        - rating
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        courseId:
          type: integer
        courseTitle:
          type: string
        courseThumbnail:
          type: string
          nullable: true
        trainerName:
          type: string
        progressPercentage:
          type: number
        completionStatus:
          type: string
          enum: [in_progress, completed]
        completedLessons:
          type: array
          items:
            type: integer
        enrollmentDate:
          type: string
      required:
        - id
        - studentId
        - courseId
        - courseTitle
        - trainerName
        - progressPercentage
        - completionStatus
        - completedLessons
        - enrollmentDate
    EnrollBody:
      type: object
      properties:
        courseId:
          type: integer
      required:
        - courseId
    UpdateProgressBody:
      type: object
      properties:
        lessonId:
          type: integer
        completed:
          type: boolean
      required:
        - lessonId
        - completed
...